ipmap/requirements.txt
decorator==4.0.4
geoip2==2.2.0
ipython==4.0.1
ipython-genutils==0.1.0
maxminddb==1.2.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
requests==2.8.1
scapy-python3==0.18
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
